+++ src/api/address.ts
@@ -1,7 +1,8 @@
 import { API_URL } from '@/constant/network'
 import { throwResponseStatusThenChaining } from '@/lib/network'
 import { serverToClientAddress } from '@/lib/utils'
-import { ClientAddress, ServerAddress, UpdateAddressDto } from '@/types/addresses'
+import { ClientAddress, UpdateAddressDto } from '@/types/addresses'
+import { BaseApiResponse } from '@/types/api'
 
 export const addressApi = {
   getAddress: async () => {
@@ -12,7 +13,7 @@ export const addressApi = {
       },
     })
       .then(throwResponseStatusThenChaining)
-      .then((res) => res.json() as Promise<ServerAddress>)
+      .then((res) => res.json() as Promise<AddressResponse>)
       .then((data) => serverToClientAddress(data))
   },
   updateAddress: async (id: number, body: Partial<ClientAddress>) => {
@@ -24,7 +25,7 @@ export const addressApi = {
       body: JSON.stringify(body),
     })
       .then(throwResponseStatusThenChaining)
-      .then((res) => res.json() as Promise<ServerAddress>)
+      .then((res) => res.json() as Promise<AddressResponse>)
       .then((data) => serverToClientAddress(data))
   },
   saveAddress: async (data: UpdateAddressDto) => {
@@ -36,7 +37,17 @@ export const addressApi = {
       body: JSON.stringify(data),
     })
       .then(throwResponseStatusThenChaining)
-      .then((res) => res.json() as Promise<ServerAddress>)
+      .then((res) => res.json() as Promise<AddressResponse>)
       .then((data) => serverToClientAddress(data))
   },
 }
+
+export type ServerAddress = {
+  deliveryAddressId: number
+  recipientName: string
+  phoneNumber: string
+  roadAddress: string
+  detailAddress: string
+  zipCode: string
+}
+export type AddressResponse = BaseApiResponse<ServerAddress>
