+++ src/types/addresses.ts
@@ -6,21 +6,6 @@ export type AddressForm = {
   address: AddressState
 }
 
-export type ServerAddress = {
-  success: boolean
-  message: string
-  result: ServerAddressField
-}
-
-export type ServerAddressField = {
-  deliveryAddressId: number
-  recipientName: string
-  phoneNumber: string
-  roadAddress: string
-  detailAddress: string
-  zipCode: string
-}
-
 export type UpdateAddressDto = {
   recipientName: string
   phoneNumber: string
