+++ src/types/api.ts
@@ -1,5 +1,5 @@
 export type BaseApiResponse<T> = {
   success: boolean
   message: string
-  result?: T
+  result: T
 }
